Add relation of disease with food consumption level
Add a new php file that loads the disease_food_consumption table.
Modified the code so the base url is based on the environment loaded into the app.
Added two examples of how a patient report looks with the new food consumption level.
Add missing files that loads additional data for the html medical reports.
Add mysqli extension to php so that files work as the files currently in production.