Add files to fill medical report
Add one single html file that reads the user id and loads all its information, this way it is no longer required to load each patient as a different html file.
Add missing files from the original project that are used in informefinal.html
Modified php files to no longer store or return the url from the DB and instead just check that the patient exists before continuing, and allow javascript to build the location.
Modified informefinal.html to informefinal.php file type to allow checking if user is logged in or not.

Note: That current architecture and logic does not protect a logged user to check another users information if logged in.